requirements.txt
markdown
google-generativeai>=0.8.5
